simplifying code changes
